requirements.txt
numpy
pickle
pytorch>=1.5.0,<=1.7.1
torchvision